svm: move a lot of CEntryPoint code to guest.staging
