+++ sdk/mx.sdk/suite.py
@@ -1173,6 +1173,7 @@ suite = {
                    org.graalvm.truffle.runtime.svm,
                    org.graalvm.nativeimage.builder,
                    org.graalvm.nativeimage.foreign,
+                   org.graalvm.nativeimage.guest.staging,
                    org.graalvm.nativeimage.shared,
                    com.oracle.graal.graal_enterprise,
                    com.oracle.svm.svm_enterprise,
