+++ substratevm/src/com.oracle.svm.guest.staging/src/com/oracle/svm/guest/staging/c/function/CEntryPointErrors.java
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.oracle.svm.core.c.function;
+package com.oracle.svm.guest.staging.c.function;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
